feat: create isAdmin controller, add error handling

+++ src/controller/notice/controller.ts
@@ -3,18 +3,23 @@
 
 import { RequestHandler } from 'express';
 import NoticeService from '../../service/notice.service';
-import { BadRequestError } from '../../util/customErrors';
+import { BadRequestError, ForbiddenError } from '../../util/customErrors';
 import { NoticeResDTO, NoticeUpdateReqDTO } from '../../type/notice.dto';
-import TeamRepository from '../../repository/team.repository';
+
+declare module 'express-session' {
+  interface SessionData {
+    passport: {
+      user: number;
+    };
+  }
+}
 
 export const getNotices: RequestHandler = async (req, res, next) => {
   try {
-    const teamId = Number(req.params.teamId); //
-    const team = await TeamRepository.findOne({
-      where: { id: teamId },
-    });
-    if (!team) throw new BadRequestError('해당하는 모임이 없습니다.');
-    const notices = await NoticeService.getNoticeById(team);
+    const teamId = Number(req.params.teamId);
+    if (isNaN(teamId)) throw new BadRequestError('teamId가 숫자가 아닙니다.');
+    const notices = await NoticeService.getNoticeByTeamId(teamId);
+    if (!notices) throw new BadRequestError('해당하는 모임이 없습니다.');
     const noticeList = notices.map((notice) => {
       return <NoticeResDTO>{
         noticeId: notice.id,
@@ -23,7 +28,6 @@ export const getNotices: RequestHandler = async (req, res, next) => {
         isPrior: notice.isPrior,
       };
     });
-
     res.status(200).json(noticeList);
   } catch (error) {
     next(error);
@@ -33,11 +37,12 @@ export const getNotices: RequestHandler = async (req, res, next) => {
 export const createNotice: RequestHandler = async (req, res, next) => {
   try {
     const teamId = Number(req.params.teamId);
-    const team = await TeamRepository.findOne({
-      where: { id: teamId },
-    });
+    const team = await NoticeService.findTeamById(teamId);
     if (!team) throw new BadRequestError('해당하는 모임이 없습니다.');
+
     const NoticeUpdateReqDTO: NoticeUpdateReqDTO = req.body;
+    if (NoticeUpdateReqDTO.startDate > NoticeUpdateReqDTO.endDate)
+      throw new BadRequestError('시작일이 끝나는 일보다 늦을 수 없습니다.');
     NoticeService.createNotice(
       team,
       NoticeUpdateReqDTO.content,
@@ -45,25 +50,21 @@ export const createNotice: RequestHandler = async (req, res, next) => {
       NoticeUpdateReqDTO.endDate,
       NoticeUpdateReqDTO.isPrior,
     );
-    //400 BadRequest 일 경우는?
-    //공지사항에 내용도 없고 시작시간, 끝 시간이 없으면 좀 그렇지 않을까? 공간만 차지할 듯
-    //그리고 teamId가 없는 teamId일때? 시작시간, 끝시간이 순서가 맞지 않을 때?
-    //사실 teamId 빼고는 프론트엔드에서 다 되긴 할 텐데
-    //401? admin이 아닐 때, admin이지만 본인이 admin인 teamId와 요청의 teamId가 일치하지 않을 때
     return res.status(200).json();
   } catch (error) {
     next(error);
   }
 };
 
-export const isAdmin: RequestHandler = async (req, res, next) => {
-  //memberId를 받지 않고 요청을 한 사람이 admin인지 알아야 함
-};
-
 export const updateNotice: RequestHandler = async (req, res, next) => {
   try {
     const noticeId = Number(req.params.noticeId);
+    const notice = await NoticeService.findNoticeById(noticeId);
+    if (!notice) throw new BadRequestError('해당하는 공지가 없습니다.');
+
     const NoticeUpdateReqDTO: NoticeUpdateReqDTO = req.body;
+    if (NoticeUpdateReqDTO.startDate > NoticeUpdateReqDTO.endDate)
+      throw new BadRequestError('시작일이 끝나는 일보다 늦을 수 없습니다.');
     NoticeService.updateNotice(
       noticeId,
       NoticeUpdateReqDTO.content,
@@ -76,11 +77,78 @@ export const updateNotice: RequestHandler = async (req, res, next) => {
     next(error);
   }
 };
+
 export const deleteNotice: RequestHandler = async (req, res, next) => {
   try {
     const noticeId = Number(req.params.noticeId);
+    const notice = await NoticeService.findNoticeById(noticeId);
+    if (!notice) throw new BadRequestError('해당하는 공지가 없습니다.');
+
     NoticeService.deleteNoticeById(noticeId);
     return res.status(200).json();
+  } catch (error) {
+    next(error);
+  }
+};
+
+export const isAdmin: RequestHandler = async (req, res, next) => {
+  try {
+    const memberId = Number(req.session.passport?.user);
+    const { teamId, noticeId, scheduleId } = req.params;
+
+    if (teamId) {
+      if (isNaN(Number(teamId)))
+        throw new BadRequestError('teamId가 숫자가 아닙니다.');
+
+      const team = await NoticeService.findTeamById(Number(teamId));
+      if (!team) throw new BadRequestError('해당하는 모임이 없습니다.');
+
+      const memberTeam = await NoticeService.findMemberTeam(
+        memberId,
+        Number(teamId),
+      );
+      if (!memberTeam)
+        throw new BadRequestError('해당하는 모임에 소속되어 있지 않습니다.');
+      if (!memberTeam.isAdmin) throw new ForbiddenError('권한이 없습니다.');
+    } else if (noticeId) {
+      if (isNaN(Number(noticeId)))
+        throw new BadRequestError('noticeId가 숫자가 아닙니다.');
+
+      const teamId = await NoticeService.getTeamIdByNoticeId(Number(noticeId));
+      if (!teamId) throw new BadRequestError('해당하는 공지가 없습니다.');
+
+      const memberTeam = await NoticeService.findMemberTeam(memberId, teamId);
+      if (!memberTeam)
+        throw new BadRequestError('해당하는 모임에 소속되어 있지 않습니다.');
+      if (!memberTeam.isAdmin) throw new ForbiddenError('권한이 없습니다.');
+    } else if (scheduleId) {
+      if (isNaN(Number(scheduleId)))
+        throw new BadRequestError('scheduleId가 숫자가 아닙니다.');
+
+      const schedule = await NoticeService.findScheById(Number(scheduleId));
+      if (!schedule) throw new BadRequestError('해당하는 일정이 없습니다.');
+
+      const teamSchedule = await NoticeService.findTeamScheByScheId(
+        Number(scheduleId),
+      );
+      if (teamSchedule) {
+        const teamId = teamSchedule.team.id;
+        const isPublic = teamSchedule.team.isPublic;
+        if (!isPublic) next();
+        else {
+          const memberTeam = await NoticeService.findMemberTeam(
+            memberId,
+            teamId,
+          );
+          if (!memberTeam)
+            throw new BadRequestError(
+              '해당하는 모임에 소속되어 있지 않습니다.',
+            );
+          if (!memberTeam.isAdmin) throw new ForbiddenError('권한이 없습니다.');
+        }
+      }
+    }
+    next();
   } catch (error) {
     next(error);
   }
